[🚚rename]: 인증 관련 type 파일 디렉토리 위치 이동
